Fixed potential breaking changes

+++ langchain4j-oracle/src/main/java/dev/langchain4j/store/embedding/oracle/EmbeddingTable.java
@@ -1,22 +1,19 @@
 package dev.langchain4j.store.embedding.oracle;
 
+import static dev.langchain4j.internal.ValidationUtils.ensureNotNull;
+
 import dev.langchain4j.data.document.Metadata;
 import dev.langchain4j.data.embedding.Embedding;
 import dev.langchain4j.data.segment.TextSegment;
 import dev.langchain4j.store.embedding.filter.Filter;
-import oracle.jdbc.OracleType;
-
-import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.sql.SQLType;
 import java.sql.Statement;
 import java.util.List;
 import java.util.UUID;
 import java.util.function.BiFunction;
-
-import static dev.langchain4j.internal.ValidationUtils.ensureNotNull;
-
+import javax.sql.DataSource;
+import oracle.jdbc.OracleType;
 
 /**
  * <p>
@@ -102,21 +99,18 @@ public final class EmbeddingTable {
 
     /**
      * Creates a table configured by the {@link Builder} of this EmbeddingTable. No table is created if the Builder was
-     * configured with {@link CreateOption#DO_NOT_CREATE}.
+     * configured with {@link CreateOption#CREATE_NONE}.
      *
      * @param dataSource Data source that connects to an Oracle Database where the table is (possibly) created.
      *
      * @throws SQLException If an error prevents the table from being created.
      */
     void create(DataSource dataSource) throws SQLException {
-        if (createOption == CreateOption.DO_NOT_CREATE)
-            return;
+        if (createOption == CreateOption.CREATE_NONE) return;
 
         try (Connection connection = dataSource.getConnection();
-             Statement statement = connection.createStatement();
-        ) {
-            if (createOption == CreateOption.CREATE_OR_REPLACE)
-                statement.addBatch("DROP TABLE IF EXISTS " + name);
+                Statement statement = connection.createStatement(); ) {
+            if (createOption == CreateOption.CREATE_OR_REPLACE) statement.addBatch("DROP TABLE IF EXISTS " + name);
 
             statement.addBatch("CREATE TABLE IF NOT EXISTS " + name
                     + "(" + idColumn + " VARCHAR(36) NOT NULL, "
@@ -149,10 +143,9 @@ public final class EmbeddingTable {
     String mapMetadataKey(String key, OracleType type) {
         // Oracle JDBC does not implement getName() correctly for BINARY_FLOAT and BINARY_DOUBLE; It puts a space where
         // the underscore should be.
-        String typeName =
-            type == OracleType.BINARY_FLOAT ? "BINARY_FLOAT"
-                : type == OracleType.BINARY_DOUBLE ? "BINARY_DOUBLE"
-                : type.getName();
+        String typeName = type == OracleType.BINARY_FLOAT
+                ? "BINARY_FLOAT"
+                : type == OracleType.BINARY_DOUBLE ? "BINARY_DOUBLE" : type.getName();
 
         return "JSON_VALUE(" + metadataColumn + ", '$." + key + "' RETURNING " + typeName + " NULL ON ERROR)";
     }
@@ -217,7 +210,7 @@ public final class EmbeddingTable {
     public static class Builder {
 
         // These fields are specified by method level JavaDocs
-        private CreateOption createOption = CreateOption.DO_NOT_CREATE;
+        private CreateOption createOption = CreateOption.CREATE_NONE;
         private String name;
         private String idColumn = "id";
         private String embeddingColumn = "embedding";
@@ -227,7 +220,7 @@ public final class EmbeddingTable {
         private Builder() {}
 
         /**
-         * Configures the option to create (or not create) a table. The default is {@link CreateOption#DO_NOT_CREATE},
+         * Configures the option to create (or not create) a table. The default is {@link CreateOption#CREATE_NONE},
          * which means that no attempt is made to create a table.
          *
          * @param createOption Option for creating the index. Not null.
@@ -319,6 +312,5 @@ public final class EmbeddingTable {
 
             return new EmbeddingTable(this);
         }
-
     }
 }
